requirements.txt
neo4j
Django
django-neomodel